+++ src/utils/jwt.ts
@@ -26,6 +26,30 @@ export const getJwtFromCookies = (
   return { jwtPayload: decodeJwt(jwt), jwt };
 };
 
+export const getJwtFromMiddlewareCookies = (req: NextRequest) => {
+  const COOKIE_SECRET = process.env.COOKIE_SECRET;
+
+  const userTokenCookie = req.cookies.get(USER_TOKEN_NAME);
+
+  if (!userTokenCookie)
+    throw new Error(
+      `Failed to verify signature for ${USER_TOKEN_NAME} cookie: cookie not found`,
+    );
+
+  const cookieValue = userTokenCookie.value;
+
+  // If the cookie is not a signed cookie, the parser will return the provided value
+  const unsignedCookie = cookieParser.signedCookie(cookieValue, COOKIE_SECRET);
+
+  if (!unsignedCookie || unsignedCookie === 'undefined') {
+    throw new Error(
+      `Failed to verify signature for ${USER_TOKEN_NAME} cookie: ${cookieValue}`,
+    );
+  }
+
+  return { jwt: unsignedCookie, jwtPayload: decodeJwt(unsignedCookie) };
+};
+
 export const axiosConfig = (jwt: string) => {
   return {
     withCredentials: true,
